refactor: Cleaned folder tree
