+++ src/lib/convertJSONAnnotationToInline.js
